feat: Add logout functionality
Add a logout endpoint and view to the authentication module. This allows users to log out of the application by sending a POST request to the logout URL. The logout view invalidates the refresh token and returns a success message upon successful logout. This feature enhances the user experience by providing a way for users to securely log out of the application.